Extract 'filtration'-related functions into separate module
