+++ lz/iterating.py
@@ -11,14 +11,11 @@ from typing import (Hashable,
                     Sequence,
                     Tuple)
 
-from .functional import (combine,
-                         compose)
+from .functional import compose
 from .hints import (Domain,
                     Map,
                     Operator,
-                    Predicate,
                     Range)
-from .replication import duplicate
 
 
 def mapper(map_: Map) -> Map[Iterable[Domain], Iterable[Range]]:
@@ -35,63 +32,6 @@ def flatmapper(map_: Map[Domain, Iterable[Range]]
     and flattens results.
     """
     return compose(flatten, mapper(map_))
-
-
-def sifter(predicate: Predicate = None) -> Operator[Iterable[Domain]]:
-    """
-    Returns function that selects elements from iterable
-    which satisfy given predicate.
-
-    If predicate is not specified than true-like objects are selected.
-    """
-    return functools.partial(filter, predicate)
-
-
-def scavenger(predicate: Predicate = None) -> Operator[Iterable[Domain]]:
-    """
-    Returns function that selects elements from iterable
-    which dissatisfy given predicate.
-
-    If predicate is not specified than false-like objects are selected.
-    """
-    return functools.partial(itertools.filterfalse, predicate)
-
-
-def separator(predicate: Predicate = None
-              ) -> Map[Iterable[Domain],
-                       Tuple[Iterable[Domain], Iterable[Domain]]]:
-    """
-    Returns function that returns pair of iterables
-    first of which consists of elements that dissatisfy given predicate
-    and second one consists of elements that satisfy given predicate.
-    """
-    return compose(tuple,
-                   combine([scavenger(predicate), sifter(predicate)]),
-                   duplicate)
-
-
-def grabber(predicate: Predicate = None) -> Operator[Iterable[Domain]]:
-    """
-    Returns function that selects elements from the beginning of iterable
-    while given predicate is satisfied.
-
-    If predicate is not specified than true-like objects are selected.
-    """
-    if predicate is None:
-        predicate = bool
-    return functools.partial(itertools.takewhile, predicate)
-
-
-def kicker(predicate: Predicate = None) -> Operator[Iterable[Domain]]:
-    """
-    Returns function that skips elements from the beginning of iterable
-    while given predicate is satisfied.
-
-    If predicate is not specified than true-like objects are skipped.
-    """
-    if predicate is None:
-        predicate = bool
-    return functools.partial(itertools.dropwhile, predicate)
 
 
 def cutter(slice_: slice) -> Operator[Iterable[Domain]]:
