+++ tests/filtration_tests/test_separator.py
@@ -2,8 +2,8 @@ from operator import not_
 from typing import (Any,
                     Iterable)
 
+from lz.filtration import separator
 from lz.hints import Predicate
-from lz.iterating import separator
 from lz.logical import negate
